20-02-2018

+++ ft_atoi.c
@@ -6,16 +6,16 @@
 /*   By: mbelalou <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/15 19:52:45 by mbelalou          #+#    #+#             */
-/*   Updated: 2018/02/12 15:51:27 by mbelalou         ###   ########.fr       */
+/*   Updated: 2018/02/13 19:15:34 by mbelalou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	char	*s;
-	int		signe;
+	long		signe;
 
 	s = (char *)str;
 	while (*s && ((*s < 33 && *s != 27) && *s != '\200'))
@@ -26,5 +26,5 @@ int		ft_atoi(const char *str)
 		signe = (*s == '-') ? -1 : 1;
 		s++;
 	}
-	return ((*s) ? signe * ft_base_to_decimal(s, DECIMAL) : 0);
+	return ((*s) ? (long)signe * ft_base_to_decimal(s, DECIMAL) : 0);
 }
